Implement TempDir in cc file
* Implemented TempDir class
  in temp-dir.cc file.
