+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tests/configuration_test.cc
@@ -20,6 +20,7 @@
 #include "common/configuration.h"
 #include "common/configuration_loader.h"
 #include "utils/temp-file.h"
+#include "utils/temp-dir.h"
 
 #include <gmock/gmock.h>
 #include <cstdio>
