+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tests/hdfs_builder_test.cc
@@ -19,6 +19,7 @@
 #include "hdfspp/hdfs_ext.h"
 #include "configuration_test.h"
 #include "utils/temp-file.h"
+#include "utils/temp-dir.h"
 
 #include <gmock/gmock.h>
 #include <google/protobuf/stubs/common.h>
